Fix Race Condition in Axios Interceptor for Refresh Token (#356)
* Fix refresh token race condition

* Change the login loading condition

* Improve variable readability

* Fix typo
