+++ frontend/src/providers/AuthProvider.tsx
@@ -9,14 +9,15 @@ interface AuthProviderProps {
 
 function AuthProvider(props: AuthProviderProps) {
 	const { children } = props;
-	const { data: user, isSuccess, isLoading } = useGetUserQuery();
+	const { data: user, isSuccess, isLoading, isPending } = useGetUserQuery();
 	const shouldChangeNickname = useMemo(
 		() => isSuccess && !user.nickname,
 		[isSuccess, user?.nickname]
 	);
+	const isAuthLoading = isLoading || isPending;
 
 	return (
-		<AuthContext.Provider value={{ isLoggedIn: isSuccess, isLoading }}>
+		<AuthContext.Provider value={{ isLoggedIn: isSuccess, isLoading: isAuthLoading }}>
 			{shouldChangeNickname ? <ChangeNicknameModal open /> : children}
 		</AuthContext.Provider>
 	);
